flask/requirements.txt
Flask==1.1.2
mysql-connector==2.2.9
flask-mysql==1.5.2
